[Refactor](Completion): Moved config to appropriate for this layer folder.
