+++ lua/config/Completion/cmp/init.lua
@@ -1,0 +1,3 @@
+return {
+  settings = require('config.Completion.cmp.settings'),
+}
