+++ lua/config/Completion/cmp/settings.lua
@@ -3,13 +3,13 @@ local snippets_engine = require('config.lsp.luasnip_engine')
 local cmp = require('cmp')
 local compare = require('cmp.config.compare')
 
-local window_visuals = require('config.lsp.completion.window_visuals')
-local formatting = require('config.lsp.completion.formatting')
-local keymappings = require('config.lsp.completion.keymappings')
+local window_visuals = require('config.Completion.cmp.window_visuals')
+local formatting = require('config.Completion.cmp.formatting')
+local keymappings = require('config.Completion.cmp.keymappings')
 
 local MIN_KEYWORD_LENGTH = 2
 
-cmp.setup({
+return {
   window = window_visuals,
 
   formatting = formatting,
@@ -85,51 +85,4 @@ cmp.setup({
     throttle = 60,
     fetching_timeout = 200,
   }
-})
-
--- * Enabling support for autopairs.
--- * If you want insert `(` after select function or method item
-local cmp_autopairs = require('nvim-autopairs.completion.cmp')
-cmp.event:on(
-  'confirm_done',
-  cmp_autopairs.on_confirm_done({ map_char = { tex = '' } })
-)
-
--- ? Is it used only for lisp?.
--- * Add a lisp filetype (wrap my-function), FYI: Hardcoded = { "clojure",
---"clojurescript", "fennel", "janet" }
---cmp_autopairs.lisp[#cmp_autopairs.lisp+1] = "racket"
-
--- Unfortunately, `= true` wasn't working.
-local autocomplete_on_every_stroke = {
-  --require('cmp.types').cmp.TriggerEvent.InsertEnter,
-  require('cmp.types').cmp.TriggerEvent.TextChanged,
 }
-
--- Use buffer source for `/`.
-cmp.setup.cmdline('/', {
-  autocomplete = autocomplete_on_every_stroke,
-  sources = {
-    { name = 'buffer', option = { keyword_pattern = [=[[^[:blank:]].*]=] } },
-  },
-})
-
--- Use cmdline & path source for ':'.
-cmp.setup.cmdline(':', {
-  completion = {
-    autocomplete = autocomplete_on_every_stroke,
-  },
-  sources = cmp.config.sources({
-    { name = 'path', keyword_length = 2 },
-  }, {
-    -- Don't show autocomplete on things like `:w`.
-    { name = 'cmdline', keyword_length = 2 },
-  }),
-
-  performance = {
-    debounce = 300,
-    throttle = 60,
-    fetching_timeout = 200,
-  }
-})
-
